fix save, save&close and remove in custom form

+++ Js v2/Forms/custom_form.js
@@ -7,12 +7,14 @@ export default class CustomForm {
     resolvePromise;
     targetData;
     flag;
+    targetId;
     constructor(){
         this.builder = null;
         this.values = [];
         this.resolvePromise = null;
         this.targetData = null;
         this.flag = false;
+        this.targetId = 0;
     }
 
     // initialize(){
@@ -28,13 +30,12 @@ export default class CustomForm {
     // }
 
     async initialize(){
+        //Get Form
         const form = await this.getForm();
         const mainForm = form[0];
         const formAfterParse = JSON.parse(mainForm.fromJson);
-
+        //Get Data
         this.targetData = JSON.parse(localStorage.getItem('targetData'));
-        this.flag = JSON.parse(localStorage.getItem('newRecordFlag'))
-        console.log('flag' , this.flag);
         this.builder = new FormBuilder(formAfterParse, 'custom' ,'form');
 
         let saveBtn = document.getElementById('save');
@@ -43,21 +44,26 @@ export default class CustomForm {
         let newBtn = document.getElementById('new');
         newBtn.addEventListener('click' ,this.handleNewBtn);
         
+        // if(this.targetData !== null){
+        //     this.builder.mapData(this.targetData);
+        //     console.log('Updated');
+        //     saveBtn.addEventListener('click', ()=> this.handleNewSaveBtn('PUT', false,this.targetData.id))
+        //     saveandcloseBtn.addEventListener('click',()=> this.handleNewSaveBtn('PUT', true, this.targetData.id))
+        //     
+        // }else{
+        //     console.log('Add');
+        //     saveBtn.addEventListener('click',()=> this.handleNewSaveBtn('POST', false));
+        //     saveandcloseBtn.addEventListener('click',()=> this.handleNewSaveBtn('POST', true));
+        // }
+
         if(this.targetData !== null){
             this.builder.mapData(this.targetData);
-            console.log('Updated');
-            saveBtn.addEventListener('click', ()=>this.handleNewSaveBtn('PUT', false,this.targetData.id))
-            saveandcloseBtn.addEventListener('click',()=>this.handleNewSaveBtn('PUT', true, this.targetData.id))
-            removeBtn.addEventListener('click' , ()=>this.handleRemoveBtn);
-        }else{
-            console.log('Add');
-            saveBtn.addEventListener('click',()=>this.handleNewSaveBtn('POST', false));
-            saveandcloseBtn.addEventListener('click',()=>this.handleNewSaveBtn('POST', true));
         }
 
-        
-        
-
+        saveBtn.addEventListener('click',()=> this.handleNewSaveBtn(false));
+        saveandcloseBtn.addEventListener('click',()=> this.handleNewSaveBtn(true));
+        removeBtn.addEventListener('click' , ()=> this.handleRemoveBtn());
+    
     }
 
     async getForm(){
@@ -93,26 +99,40 @@ export default class CustomForm {
     //     }
     // }
 
-    async handleNewSaveBtn(method,shouldClose = false,id=''){
+    async handleNewSaveBtn(shouldClose){
         let dataObject = {}
         for(let i=0; i< this.builder.Fields.length; i++){
             let key = this.builder.Fields[i].name;
             let value = document.getElementById(this.builder.Fields[i].id).value
             dataObject[key]  = value;
         }
-        try{
 
-            method = this.flag===true ? 'PUT':'POST';
-            // if(method === 'PUT'){
-            //     let data = localStorage.getItem() 
-            // }
-            const response = await this.pushDataIntoDB(dataObject , method , id);
-            console.log('response', response);
-            // localStorage.setItem('newRecordFlag', true);
+        try{
+            let response = null;
+            let dataFromLocalStorage = localStorage.getItem('newRecordFlag');
+            let data = JSON.parse(dataFromLocalStorage);
+            console.log('data', data);
+            console.log('dataFromLocalStorage:', dataFromLocalStorage);
+            if(this.targetData !== null || data !== null){
+                console.log('update');
+                this.targetId = this.targetData?.id || data?.id;
+                console.log('id in localStorage: ' , id);
+                response = await this.pushDataIntoDB(dataObject , 'PUT' , this.targetId);
+            }
+            else{
+                console.log('Add')
+                response = await this.pushDataIntoDB(dataObject , 'POST');
+                localStorage.setItem('newRecordFlag', JSON.stringify(response))
+                console.log('responsesdf', response);
+            }
+            // console.log('response', response);
+
             if (shouldClose) {
-                // window.close();
-                // window.location.reload();
-                // window.history.back();
+                localStorage.setItem('targetData', null);
+                localStorage.setItem('newRecordFlag',null);
+                window.close();
+                window.location.reload();
+                window.history.back();
             }
         }catch(error){
             console.error('Error pushing data into DB:', error);
@@ -120,25 +140,32 @@ export default class CustomForm {
     }
 
     handleRemoveBtn(){
-        const id = this.targetData.id;
-        let isDelete = alert('Are you sure you want to Delete this record?');
-        if(isDelete){
-            const response = fetch(`http://localhost:5032/api/Employees/${id}`,{
-                method:'DELETE',
-                headers:{
-                    'Content-Type': 'application/json'
-                }
-            });
+        const id = this.targetId;
+        if(id !== 0 ){
+            let isDelete = confirm('Are you sure you want to Delete this record?');
+            if(isDelete){
+                const response = fetch(`http://localhost:5032/api/Employees/${id}`,{
+                    method:'DELETE',
+                    headers:{
+                        'Content-Type': 'application/json'
+                    }
+                });
+            }
+        }else{
+            alert('There is no record to delete')
         }
+        
         
     }
 
     handleNewBtn(){
         localStorage.setItem('targetData', null);
+        localStorage.setItem('newRecordFlag',null);
         window.location.reload();
     }
 
-    async pushDataIntoDB(data , method , id){
+    async pushDataIntoDB(data , method , id=''){
+        console.log('id', id);
         data.departmentId = 1;
         console.log('data: ', data);
         const response = await fetch(`http://localhost:5032/api/Employees/${id}`,{
@@ -148,7 +175,9 @@ export default class CustomForm {
             },
             body: JSON.stringify(data)
         });
-        return response;
+        if(method === 'POST'){
+            return await response?.json();
+        }
     }
